Add files via upload
Design and code the items in a library including books, magazines, video DVDs and CDs.

include an abstract superclass and/or interface for the common information a library might have for each item, such as item number and title and functionality such as reserve, checkIn, checkOut, and payFines. Add additional instance variables or methods as needed to model the characteristics of each class. 
In addition, each class should implement a Printable interface that provides a way to print a summary of the item. 

Add a client to test your classes.